pyproject.toml
[tool.poetry]
name = "django_app"
version = "0.1.0"
description = ""
authors = ["ivanpobeguts <ivpobeguts@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"